fix: keep main.rs and run tests under the 500-line limit
Extract the config-render block from main.rs into startup::render_config, and
move the new run flag-parity tests into their own engine_integration/run_flags
module.
