+++ tests/engine_integration/run_flags.rs
@@ -1,0 +1,157 @@
+//! `run` flag-parity integration tests (split for the source line limit).
+use super::*;
+
+#[tokio::test]
+async fn engine_run_applies_user_workdir_env_and_entrypoint() {
+	let client = match podman().await {
+		Some(d) => d,
+		None => return,
+	};
+	let proj = proj("ruwe");
+	let engine = Engine::new(client, proj.clone());
+	let file = parse_str("services:\n  job:\n    image: alpine:latest\n").unwrap();
+
+	// --entrypoint sh, cmd is its args; -u root, -w /tmp, -e MARK=ok. The command
+	// only exits 0 when the working dir, user id and env override all took effect.
+	let result = engine
+		.run(
+			&file,
+			"job",
+			podup::RunOptions {
+				cmd: vec![
+					"-c".to_string(),
+					"test \"$(pwd)\" = /tmp && test \"$(id -u)\" = 0 && test \"$MARK\" = ok"
+						.to_string(),
+				],
+				rm: true,
+				entrypoint: Some("sh".to_string()),
+				user: Some("0".to_string()),
+				workdir: Some("/tmp".to_string()),
+				env_overrides: vec!["MARK=ok".to_string()],
+				..Default::default()
+			},
+		)
+		.await;
+	assert!(
+		result.is_ok(),
+		"run with user/workdir/env/entrypoint failed: {result:?}"
+	);
+}
+
+#[tokio::test]
+async fn engine_run_applies_volume_publish_and_interactive() {
+	let client = match podman().await {
+		Some(d) => d,
+		None => return,
+	};
+	let dir = tempfile::tempdir().unwrap();
+	fs::write(dir.path().join("marker.txt"), b"present").unwrap();
+	let mount = format!("{}:/mnt/in:ro", dir.path().to_str().unwrap());
+
+	let proj = proj("rvp");
+	let engine = Engine::new(client, proj.clone());
+	let file = parse_str("services:\n  job:\n    image: alpine:latest\n").unwrap();
+
+	// -v bind-mounts the host dir, -i keeps stdin open, -p publishes an ad-hoc
+	// port. The command only exits 0 if the mounted marker is readable.
+	let result = engine
+		.run(
+			&file,
+			"job",
+			podup::RunOptions {
+				cmd: vec![
+					"sh".to_string(),
+					"-c".to_string(),
+					"test \"$(cat /mnt/in/marker.txt)\" = present".to_string(),
+				],
+				rm: true,
+				volumes: vec![mount],
+				publish: vec!["127.0.0.1:0:9".to_string()],
+				interactive: true,
+				..Default::default()
+			},
+		)
+		.await;
+	assert!(
+		result.is_ok(),
+		"run with volume/publish/interactive failed: {result:?}"
+	);
+}
+
+#[cfg(feature = "test-helpers")]
+#[tokio::test]
+async fn engine_run_no_deps_skips_dependency_startup() {
+	let client = match podman().await {
+		Some(d) => d,
+		None => return,
+	};
+	let proj = proj("rnd");
+	let engine = Engine::new(client, proj.clone());
+	let file = parse_str(
+		"services:\n  job:\n    image: alpine:latest\n    depends_on:\n      - dep\n  dep:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	// With --no-deps the dependency must not be started.
+	let result = engine
+		.run(
+			&file,
+			"job",
+			podup::RunOptions {
+				cmd: vec!["true".to_string()],
+				rm: true,
+				no_deps: true,
+				..Default::default()
+			},
+		)
+		.await;
+	let names = engine
+		.test_project_container_names()
+		.await
+		.unwrap_or_default();
+	let dep_present = names.iter().any(|n| n.contains("-dep"));
+	engine.down(&file).await.unwrap();
+	assert!(result.is_ok(), "run --no-deps failed: {result:?}");
+	assert!(
+		!dep_present,
+		"dependency container created despite --no-deps: {names:?}"
+	);
+}
+
+#[cfg(feature = "test-helpers")]
+#[tokio::test]
+async fn engine_run_starts_dependencies_by_default() {
+	let client = match podman().await {
+		Some(d) => d,
+		None => return,
+	};
+	let proj = proj("rwd");
+	let engine = Engine::new(client, proj.clone());
+	let file = parse_str(
+		"services:\n  job:\n    image: alpine:latest\n    depends_on:\n      - dep\n  dep:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	let result = engine
+		.run(
+			&file,
+			"job",
+			podup::RunOptions {
+				cmd: vec!["true".to_string()],
+				rm: true,
+				..Default::default()
+			},
+		)
+		.await;
+	let names = engine
+		.test_project_container_names()
+		.await
+		.unwrap_or_default();
+	let dep_present = names.iter().any(|n| n.contains("-dep"));
+	engine.down(&file).await.unwrap();
+	assert!(result.is_ok(), "run with deps failed: {result:?}");
+	assert!(
+		dep_present,
+		"dependency was not started by default: {names:?}"
+	);
+}
